Compare pin/release to copy
Analyse performance cost of the extra FFI call to release a pinnable slice versus the extra copy into a foreign memory segment allocated on the Java side. 1k-4k break even.

+++ java/src/main/java/org/rocksdb/FFIDB.java
@@ -281,7 +281,7 @@ public class FFIDB implements AutoCloseable {
     } catch (final Throwable methodException) {
       throw new RocksDBException(INVOCATION_ERROR + FFIMethod.GetOutput, methodException);
     }
-    final Status.Code code = Status.Code.values()[(Integer) result];
+    final Status.Code code = Status.Code.values()[result];
     switch (code) {
       case NotFound -> { return new GetOutputSlice(code, Optional.empty()); }
       case Ok -> {
